+++ plugins/pigeon/redPacket/config.d.ts
@@ -1,0 +1,12 @@
+interface redPacketData {
+  redpackets: redPacket[]
+}
+
+interface redPacket {
+  id: number
+  send_user_id: number
+  redPacket_num: number
+  pigeon_num: number
+  code: string
+  picked_user: string
+}