+++ plugins/pigeon/redPacket/index.ts
@@ -1,9 +1,11 @@
-import { eventReg } from '../../libs/eventReg.js'
-import { add, reduce } from '../pigeon/index.js'
-import { missingParams } from '../../libs/eventReg.js'
-import { getRangeCode, randomInt } from '../../libs/random.js'
-import { replyMsg } from '../../libs/sendMsg.js'
-import { getUserName } from '../../libs/Api.js'
+import type { CQEvent } from '@huan_kong/go-cqwebsocket'
+import type { commandFormat } from '@/libs/eventReg.ts'
+import { add, reduce } from '../pigeon/index.ts'
+import { eventReg } from '@/libs/eventReg.ts'
+import { missingParams } from '@/libs/eventReg.ts'
+import { getRangeCode, randomInt } from '@/libs/random.ts'
+import { replyMsg } from '@/libs/sendMsg.ts'
+import { jsonc } from 'jsonc'
 
 export default async () => {
   event()
@@ -14,29 +16,26 @@ export default async () => {
 
 //注册事件
 function event() {
-  eventReg('message', async (event, context, tags) => {
+  eventReg('message', async ({ context }, command) => {
     await get(context)
 
-    const { command } = context
+    if (!command) return
 
-    if (command) {
-      if (command.name === '鸽子红包') {
-        await give(context)
-      } else if (command.name === '剩余红包') {
-        await getAll(context)
-      }
+    if (command.name === '鸽子红包') {
+      await give(context, command)
+    } else if (command.name === '剩余红包') {
+      await getAll(context)
     }
   })
 }
 
 //我的鸽子
-async function give(context) {
-  const {
-    user_id,
-    command: { params }
-  } = context
+async function give(context: CQEvent<'message'>['context'], command: commandFormat) {
+  const { user_id } = context
 
-  if (await missingParams(context, 2)) return
+  if (await missingParams(context, command, 2)) return
+
+  const { params } = command
 
   //发送的红包数
   const redPacket_num = parseInt(params[0])
@@ -45,29 +44,23 @@ async function give(context) {
   //口令
   const code = params[2] ?? getRangeCode()
 
-  if (redPacket_num <= 0 || pigeon_num <= 0) {
+  if (redPacket_num <= 0 || pigeon_num <= 0)
     return await replyMsg(context, '红包发送失败,红包数量和鸽子数都不能<=0', { reply: true })
-  }
 
   const item = await database.select('*').where({ code }).from('red_packet').first()
-  if (item) {
-    return await replyMsg(context, '红包发送失败,该口令已存在', { reply: true })
-  }
+  if (item) return await replyMsg(context, '红包发送失败,该口令已存在', { reply: true })
 
   //校验合理性
   const pre = pigeon_num / redPacket_num
 
-  if (pre < 1) {
+  if (pre < 1)
     return await replyMsg(context, '红包发送失败,每个包需要至少一只鸽子', { reply: true })
-  }
 
-  if (parseInt(pre) !== pre) {
+  if (Math.floor(pre) !== pre)
     return await replyMsg(context, '红包发送失败,每个包里的鸽子数需要为整数', { reply: true })
-  }
 
-  if (!(await reduce({ user_id, number: pigeon_num, reason: `发送鸽子红包_${code}` }))) {
+  if (!(await reduce(user_id, pigeon_num, `发送鸽子红包_${code}`)))
     return await replyMsg(context, '红包发送失败,账户鸽子不足', { reply: true })
-  }
 
   //插入红包
   await database
@@ -85,22 +78,21 @@ async function give(context) {
   await replyMsg(context, `富哥发红包了!口令:${code}`)
 }
 
-async function get(context) {
+async function get(context: CQEvent<'message'>['context']) {
   const { user_id, message } = context
-  const { redPacketData } = global.data
-  for (let i = 0; i < redPacketData.length; i++) {
-    const item = redPacketData[i]
-    let { pigeon_num, redPacket_num, picked_user } = item
+  const { redPacketData } = global.data as { redPacketData: redPacketData }
 
-    picked_user = JSON.parse(picked_user)
+  for (let i = 0; i < redPacketData.redpackets.length; i++) {
+    const item = redPacketData.redpackets[i]
+    const { pigeon_num, redPacket_num, picked_user } = item
+
+    const pickedUserJSON = jsonc.parse(picked_user)
 
     //口令不正确
-    if (message !== item.code) {
-      continue
-    }
+    if (message !== item.code) continue
 
     //领取过
-    if (picked_user.indexOf(user_id) !== -1) {
+    if (pickedUserJSON.indexOf(user_id) !== -1) {
       await replyMsg(context, '红包领取过了哦,不要贪心啦~', { reply: true })
       continue
     }
@@ -109,15 +101,15 @@ async function get(context) {
     const getPigeonNum =
       redPacket_num === 1 ? pigeon_num : randomInt(1, (pigeon_num * randomInt(50, 70)) / 100)
 
-    picked_user.push(user_id)
+    pickedUserJSON.push(user_id)
 
-    await add({ user_id, number: getPigeonNum, reason: `领取鸽子红包_${item.code}` })
+    await add(user_id, getPigeonNum, `领取鸽子红包_${item.code}`)
 
     await database
       .update({
         redPacket_num: redPacket_num - 1,
         pigeon_num: pigeon_num - getPigeonNum,
-        picked_user: JSON.stringify(picked_user)
+        picked_user: JSON.stringify(pickedUserJSON)
       })
       .where('id', item.id)
       .from('red_packet')
@@ -126,11 +118,12 @@ async function get(context) {
       reply: true
     })
   }
+
   //更新红包列表
   await freshRedPacketList()
 }
 
-async function getAll(context) {
+async function getAll(context: CQEvent<'message'>['context']) {
   const { redPacketData } = global.data
   if (redPacketData.length !== 0) {
     let msg = ['剩余红包:']
@@ -138,9 +131,15 @@ async function getAll(context) {
     for (let i = 0; i < redPacketData.length; i++) {
       const item = redPacketData[i]
       msg.push(
-        `由${await getUserName(item.send_user_id)}发送的口令为:"${item.code}"，剩余${
-          item.pigeon_num
-        }只鸽子的红包`
+        [
+          '由',
+          await bot.get_stranger_info(item.send_user_id).then(res => res.nickname),
+          '发送的口令: ',
+          item.code,
+          ' ,剩余',
+          item.pigeon_num,
+          '只鸽子'
+        ].join()
       )
     }
 
@@ -151,7 +150,8 @@ async function getAll(context) {
 }
 
 async function freshRedPacketList() {
-  global.data.redPacketData = await database
+  const { redPacketData } = global.data as { redPacketData: redPacketData }
+  redPacketData.redpackets = await database
     .select('*')
     .where('pigeon_num', '>', 0)
     .from('red_packet')